[04]: update errors handling

+++ src/features/blogs/application/blogs.service.ts
@@ -1,7 +1,8 @@
 import { WithId } from 'mongodb'
-import { Blog } from '../types/blog.js'
+import { Blog, BlogErrorCode } from '../types/blog.js'
 import { blogsRepository } from '../repositories/blogs.repository.js'
 import { BlogInputModel } from '../models/BlogInputModel.js'
+import { DomainException } from '../../../core/exeptions/domain.exception.js'
 
 export const blogsService = {
   async findMany(): Promise<WithId<Blog>[]> {
@@ -25,6 +26,15 @@ export const blogsService = {
   },
   async delete(id: string) {
     // TODO: при удалении блога с постами - скорее всего написать, что нельзя удалять блог - DomainError
+    const isBlogHasPosts = false // тут будет запрос за постами блога и проверка
+
+    if (isBlogHasPosts) {
+      throw new DomainException(
+        'Blog has posts. Remove all posts for delete blog',
+        BlogErrorCode.HasPosts
+      )
+    }
+
     await blogsRepository.delete(id)
     return
   },
